… ecc/bls12-377/fr/sis/sis_fft.go
… ecc/bls12-377/fr/sis/sis_test.go
… ecc/bls12-377/fr/sis/test_cases.json
… ecc/bls12-381/fr/fft/fft.go
… ecc/bls12-381/fr/fft/options.go
… ecc/bls24-315/fr/fft/fft.go
… ecc/bls24-315/fr/fft/options.go
… ecc/bls24-317/fr/fft/fft.go
… ecc/bls24-317/fr/fft/options.go
… ecc/bn254/fr/fft/fft.go
… ecc/bn254/fr/fft/options.go
… ecc/bn254/fr/sis/sis.go
… ecc/bn254/fr/sis/sis.sage
… ecc/bn254/fr/sis/sis_fft.go
… ecc/bn254/fr/sis/sis_test.go
… ecc/bn254/fr/sis/test_cases.json
… ecc/bn254/fr/tensor-commitment/commitment.go
… ecc/bn254/fr/tensor-commitment/commitment_test.go
… ecc/bw6-633/fr/fft/fft.go
… ecc/bw6-633/fr/fft/options.go
… ecc/bw6-761/fr/fft/fft.go
… ecc/bw6-761/fr/fft/options.go
… field/babybear/element_arm64.s
… field/babybear/fft/fft.go
… field/babybear/fft/options.go
… field/babybear/sis/sis.go
… field/babybear/sis/sis_fft.go
… field/babybear/sis/sis_test.go
… field/babybear/vector_arm64.go
… field/generator/generator_sis.go