@gbotrel 2025-01-17
> Looks good, not very important but why didn't you use the `iter` package ?

`iter` is in go std since go 1.23 only, that would break our "compatible with latest 2 golang" promise. Syntactic sugar for later :) --> [see here in comments](https://github.com/Consensys/gnark-crypto/blob/9db9fd8fdf617a449c2e59acf4734ad97a89cbc7/field/koalabear/sis/sis.go#L212)